feat: add assets trait
